Shiny widgets - Initial test for ipyleaflet map
